[feat] cookie all ranking

+++ app/src/main/java/com/example/fitness/dto/ranking/RankingItem.kt
@@ -3,7 +3,7 @@ package com.example.fitness.dto.ranking
 data class RankingItem(
     val userName: String, // 사용자 이름
     val rank: Int,        // 사용자 랭킹
-    val currentCookieId: Int,        // 사용자 점수
+    val currentCookieId: Long,
     val totalDistance: Float,
     val consecutiveDays: Int,
     val successStreak: Boolean
